# Improve WeakList performance during staggered removals