… osu.Framework/Lists/WeakList_ValidItemsEnumerator.cs