@peppy 2020-09-23
Is the optimised case common enough to warrant 5-18% decrease across all other benchmarks? Probably worth touching on that for the record.