@smoogipoo 2020-09-23
I made a few more optimisations that bring it closer, and updated the benchmarks. The difference is coming from the additional size of `InvalidatableWeakReference`, but I don't know how to do anything about that besides not making this change in the first place...

On a profiling of loading of Centipede (the non-optimised version from earlier today), which creates all those bindables, I see the following:
![image](https://user-images.githubusercontent.com/1329837/94040547-0d9def80-fe04-11ea-9df6-810b42a462f4.png)
Which means assuming a 10% difference (`Add`), master should be getting ~690ms instead of 762ms in the same scenario - a 100ms difference. I think this is pretty irrelevant considering the sheer load this map's putting on and compared to the >1s improvement on disposal with these changes.