@smoogipoo 2020-09-24
I'd also recommend testing with https://osu.ppy.sh/beatmapsets/121591#osu/311269, for me exiting goes from 10s -> 1s after this change.