@bdach 2020-09-24
I think the code looks good now and can reproduce performance improvements.

I can get behind the enabling of nullable reference types done in these files in the last commit, but I'm not entirely sure we'll all be on the same page in that matter. @peppy is that ok by you?